DRF-Ecommerce/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
dj-database-url==0.5.0
Django==4.0.1
django-cleanup==5.2.0
django-countries==7.2.1
django-heroku==0.0.0
django-phonenumber-field==6.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
phonenumbers==8.12.41
Pillow==9.0.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-dateutil==1.5
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
waitress==2.0.0
whitenoise==5.3.0
